Election results and longevity of politicians

This is a dataset corresponding to the paper "Longevity Returns to Political Office" by Barfort, Klemmensen & Larsen (2019). The purpose of this study was to find out whether and how winning an election influences the lifespan of politicians. To answer this question, the authors collected data on all U.S. gubernatorial elections from 1945 to 2012. In addition to the names and votes of the running candidates, the data includes information on their pre- and post-election life spans. See also the 'Details' section below.

governors

Format

A tibble with 1,092 observations and 11 variables:

state

character variable indicating the state in which an election took place

year

integer variable indicating the year in which an election took place

first_name

character variable indicating the first name of a candidate

last_name

character variable indicating the last name of a candidate

party

character variable indicating a member's party

sex

character variable indicating a member's sex

died

date variable indicating the candidate's date of death

status

character variable indicating whether a candidate was the challenger or incumbent

win_margin

double variable indicating the percentage margin by which the election was won (positive values) or lost (negative values)

alive_post

integer variable indicating the number of days a candidate lived after the election took place

alive_pre

integer variable indicating the number of days a candidate lived before the election took place

Source

https://doi.org/10.7910/DVN/IBKYRX

Details

To facilitate subsequent analysis, the raw data collected by the authors was edited in three ways. First, for a given election, only the two candidates who received the highest number of votes were included. Second, candidates with unknown dates of death were excluded, resulting in fewer observations for elections in recent years. Third, in a few instances, only the year of birth or death could be determined; in these cases, the date was taken to be July 1 of that year.

Table: Data summary

Name	governors

Number of rows	1092

Number of columns	11

_______________________

Column type frequency:

Date	1

character	6

numeric	4

________________________

Group variables	None

Variable type: Date

skim_variable	n_missing	complete_rate	min	max	median	n_unique

died	0	1	1946-12-21	2019-08-18	1996-07-20	683

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

state	0	1	4	14	0	50	0

first_name	0	1	2	11	0	299	0

last_name	0	1	3	11	0	615	0

party	0	1	8	11	0	3	0

sex	0	1	4	6	0	2	0

status	0	1	9	10	0	2	0

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

year	0	1	1964.85	13.38	1945.00	1954.00	1962.00	1974.00	2011	▇▆▃▂▁

win_margin	0	1	3.22	26.07	-82.43	-9.87	0.97	12.18	100	▁▃▇▁▁

alive_post	0	1	10309.56	4885.95	46.00	6418.75	10811.00	14123.75	22067	▃▆▇▆▂

alive_pre	0	1	18892.42	3181.47	11450.00	16561.00	18757.50	20995.75	30633	▂▇▆▂▁

Author

David Kane