Kenya voter registration experiment

This dataset is from a study entitled "Electoral Administration in Fledgling Democracies:Experimental Evidence from Kenya". The study worked with Kenya's electoral commission in 1,674 communities by assigning individuals to either a control/status quo group or one of several interventions to encourage voter registration: SMS reminders, a local administrator at a polling station, canvassing, a local admin and canvassing, or a local admin and SMS.

kenya

Format

A tibble with 1,989,680 observations and 11 variables:

treatment

factor variable that indicates either the control group or some combination of treatment factors: canvassing, SMS reminder, local administrator, local admin and SMS, local admin and canvassing

poverty

a double variable for the poverty rate of the voting block for those assigned to it

distance

a double variable for average distance from polling location for those assigned to it

pop_density

a double variable for the population density of individual's block

block

a character variable for the individual's voting block ID

poll_station

a character variable for the individual's polling station ID

rv13

a double variable for the number of registered voters at polling location during 2013

reg_byrv13

a double variable for registered voters at polling location during intervention period divided by registered voters at polling location during 2013

Source

https://pdfs.semanticscholar.org/be58/c86f2366d5281d4800411d2dfcecd1a5c16a.pdf?_ga=2.260073289.590691361.1594131808-281689301.1594131808

Details

Table: Data summary

Name	kenya

Number of rows	1672

Number of columns	9

_______________________

Column type frequency:

character	2

factor	1

numeric	6

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

block	0	1	7	12	0	279	0

poll_station	0	1	7	7	0	1672	0

Variable type: factor

skim_variable	n_missing	complete_rate	ordered	n_unique	top_counts

treatment	0	1	FALSE	6	can: 279, loc: 279, loc: 279, loc: 279

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

poverty	0	1	0.43	0.11	0.18	0.35	0.43	0.49	0.90	▃▇▅▁▁

distance	0	1	20.93	16.24	0.04	10.20	17.26	27.34	179.02	▇▁▁▁▁

pop_density	0	1	0.00	0.00	0.00	0.00	0.00	0.00	0.02	▇▁▁▁▁

mean_age	8	1	41.63	2.78	24.02	39.82	41.59	43.42	50.07	▁▁▂▇▂

reg_byrv13	4	1	0.02	0.04	0.00	0.00	0.00	0.02	0.73	▇▁▁▁▁

rv13	0	1	539.23	429.35	0.00	273.00	442.00	675.00	6764.00	▇▁▁▁▁

Author

David Kane