American National Election Studies

This dataset is from the American National Election Studies (ANES), a survey that aims to provide insights into voter behavior. It has been conducted since 1948 before and after each presidential election, and combines questions about voters' political attitudes with extensive biographical information. Current as of 07/29/2020.

nes

Format

A tibble with 38,558 observations and 10 variables:

year

integer variable for year of study

state

character variable for the 2-letter abbreviation of the state of the interview of a respondent

gender

character variable with values "Male", "Female", and "Other"

income

ordered factor variable for respondents' income percentile group. Has 5 levels: 0-16th percentile, 17-33rd, 34-67th, 68-95th, 96-100th

age

ordered factor variable of respondents' age ranges, 75+ is accurate except for year 1954

education

ordered factor variable with a 7 tier delineation of educational achievement

race

character variable for respondants' race / ethnicity identification

ideology

integer variable for party identification from -3 to 3. Lower values correspond to Democrats, higher values correspond to Republicans

pres_appr

character variable of respondents' self-reported approval of the sitting president. Question was not asked before 1972

voted

character variable indicating whether the respondent voted in the national elections

region

factor variable indicating the region corresponding to the voter's survey state

Source

https://electionstudies.org/data-center/anes-time-series-cumulative-data-file/

Details

Some of the questions aksed in the survey have changed slightly over time. The ANES codebook provides further information on this issue.

Table: Data summary

Name	nes

Number of rows	38558

Number of columns	11

_______________________

Column type frequency:

character	5

factor	4

numeric	2

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

state	110	1.00	2	2	0	50	0

gender	141	1.00	4	6	0	3	0

race	287	0.99	5	15	0	6	0

pres_appr	9646	0.75	6	10	0	3	0

voted	4078	0.89	2	3	0	2	0

Variable type: factor

skim_variable	n_missing	complete_rate	ordered	n_unique	top_counts

income	2517	0.93	TRUE	5	34 : 11740, 68 : 9974, 0 -: 6300, 17 : 6213

age	450	0.99	TRUE	7	25 : 7669, 35 : 7342, 45 : 6545, 55 : 6021

education	397	0.99	TRUE	7	Hig: 9269, Som: 8540, Col: 5540, Som: 4688

region	0	1.00	FALSE	4	Sou: 13680, Mid: 10072, Wes: 7428, Nor: 7378

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

year	0	1.00	1989.40	20.50	1952	1972	1992	2012	2016	▃▃▃▃▇

ideology	619	0.98	-0.38	2.09	-3	-2	-1	2	3	▇▂▂▂▅

Author

David Kane