Student course evaluations from Harvard

Data for student feedback on Harvard courses. After each semester, the office of the Registrar at Harvard sends all students the so-called Q Guide, where students submit both quantitative and qualitative data about their courses. The primary observations of our data set are the overall rating of the course, as well as the workload.

qscores

Format

A tibble with 748 observations and 8 variables:

name

character variable with name of course

department

character variable with course department

number

character variable with course number

term

character variable with values "2018-Fall" and "2019-Spring"

enrollment

integer variable with course enrollment

hours

numeric variable representing hours per week of workload, outside of class time

rating

numeric variable representing average of students' rating of course (1 to 5 scale), with higher numbers indicating a higher rating

instructor

character variable with name of instructor

Source

The data for this tibble was generously provided by Aurash Vatan '23.

Details

The data source — skimmed from the Q Guide web app — does not support data for courses that are primarily taught in individual sections. As such, large courses like freshman Expository Writing (taught in 15-person sections) do not have data available. The data set has also been filtered to only include courses with more than 15 registered students.

Table: Data summary

Name	qscores

Number of rows	748

Number of columns	8

_______________________

Column type frequency:

character	5

numeric	3

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

name	0	1	4	97	0	687	0

department	0	1	2	8	0	86	0

number	0	1	1	6	0	504	0

term	0	1	9	11	0	2	0

instructor	0	1	5	32	0	509	0

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

enrollment	0	1	50.82	70.38	16	21.00	29.0	49.0	744.0	▇▁▁▁▁

hours	0	1	6.17	3.48	1	3.77	5.2	7.6	22.5	▇▅▂▁▁

rating	0	1	4.18	0.52	1	3.90	4.2	4.5	5.0	▁▁▁▇▇

Author

David Kane