Michigan voting experiment with social shaming

This is a dataset corresponding to the paper "Social Pressure and Voter Turnout: Evidence from a Large-Scale Field Experiment" by Gerber, Green, and Larimer (2008). See also the 'Details' section below. The aim of the study was to find out whether and to what extent people are motivated to vote by social pressure. To answer this question, the authors conducted a field experiment prior to the August 2006 primary election in Michigan. A total of 180,000 households were randomly assigned to either a control group or one of four treatment groups.

shaming

Format

A tibble with 344,084 observations and 10 variables:

sex

character variable with values "male" and "female"

birth_year

integer variable indicating the respondent birth year

primary_02

character variable indicating whether the respondent voted in the 2002 primary election

general_02

character variable indicating whether the respondent voted in the 2002 general election

primary_04

character variable indicating whether the respondent voted in the 2004 primary election

general_04

character variable indicating whether the respondent voted in the 2004 general election

treatment

factor variable indicating which of the 5 treatments were employed in 2006, but before the primary election that year: 'Control', 'Civic Duty', 'Hawthorne', 'Self', or 'Neighbors'

primary_06

0/1 integer variable indicating whether the respondent voted in the 2006 primary election

hh_size

integer variable indicating the respondent household size

no_of_names

integer variable indicating the number of names listed on the letter if the respondent was in the "Neighbours" group

Source

https://doi.org/10.1017/S000305540808009X

Details

The control group consisted of approximately 100,000 households and was observed without further intervention.The treatment groups consisted of about 20,000 households each, and were sent one mailing 11 days prior to the primary election. The first treatment group, named “Civic Duty”, received a letter that only reminded them to "do their civic duty and vote". The second treatment group, named "Hawthorne", received the same message with an additional notice that they are being studied by researchers. The letter sent to the third group, named "Self", included the content in the Hawthorne letter, but added a notice that every household member would be notified of each others' voting behaviour after the election (this information is public). The last group, "Neighbors", finally listed not only the household's voting records but also the voting records of those nearby. As in the "Self" group, everyone on the list would be notified of their voting behaviour after the primary.

Table: Data summary

Name	shaming

Number of rows	344084

Number of columns	10

_______________________

Column type frequency:

character	5

factor	1

numeric	4

________________________

Group variables	None

Variable type: character

skim_variable	n_missing	complete_rate	min	max	empty	n_unique	whitespace

sex	0	1	4	6	0	2	0

primary_02	0	1	2	3	0	2	0

general_02	0	1	2	3	0	2	0

primary_04	0	1	2	3	0	2	0

general_04	0	1	3	3	0	1	0

Variable type: factor

skim_variable	n_missing	complete_rate	ordered	n_unique	top_counts

treatment	0	1	FALSE	5	Con: 191243, Civ: 38218, Sel: 38218, Haw: 38204

Variable type: numeric

skim_variable	n_missing	complete_rate	mean	sd	p0	p25	p50	p75	p100	hist

birth_year	0	1.00	1956.21	14.45	1900	1947	1956	1965	1986	▁▁▅▇▃

primary_06	0	1.00	0.32	0.46	0	0	0	1	1	▇▁▁▁▃

hh_size	0	1.00	2.18	0.79	1	2	2	2	8	▇▂▁▁▁

no_of_names	305883	0.11	20.93	0.52	10	21	21	21	21	▁▁▁▁▇

Author

David Kane